move serialization helpers next to their tests
